packages/olympus/pyproject.toml
[project]
name = "olympus"
version = "0.3.0"
description = "Olympus agent framework — LangGraph runtime with YAML agents and append-only run logs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "chonkie>=1.6.1",
    "chromadb>=1.5.5",
    "gitpython>=3.1.46",
    "langgraph>=0.2.0",
    "magika>=1.0.2",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "sentence-transformers>=5.3.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-language-pack>=1.4.0",
    "tree-sitter-python>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
olympus = "olympus.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
